Another 32-bit fix.

+++ src/command/restore/file.c
@@ -120,7 +120,7 @@ List *restoreFile(
 
                                     // Generate delta map if block incremental
                                     if (file->blockIncrMapSize != 0)
-                                        ioFilterGroupAdd(ioReadFilterGroup(read), deltaMapNew(file->blockIncrSize));
+                                        ioFilterGroupAdd(ioReadFilterGroup(read), deltaMapNew((size_t)file->blockIncrSize));
 
                                     ioReadDrain(read);
                                 }
